JCOR-275 Added a project with test helpers
First, `withLoggedIn` has been recreated, but requires Guice (as does
the rest of the library). Now, instead of `withLoggedIn(config)(userId)`,
you can call `withLoggedIn[Env](userId)`, which should make the migration
of existing tests relatively painless.

Then, I added mock action builders that simple hold a user (or optional user),
and allow passthrough authentication/authorization with a static logged-in user.
These can be used when for testing controllers without the use of a router
(which should be preferred so all components can be mocked out).